Extract end of service report answers into separate component
This pulls things out of the SP check-answers page and into a presenter
/ view / partial which we’ll use in the PP view end of service report
page.

This follows the same pattern that George introduced for the session
feedback pages. The only new thing is that this also introduces a view
which is related to the partial, which is constructed by the containing
page’s view.

This also fixes an issue with some hardcoded IDs in the presenter tests,
which caused these tests to start to fail once I extracted this code.
